+++ HousecatEngine/src/game/CollisionMap.cpp
@@ -6,7 +6,7 @@ std::vector<std::vector<bool>> CollisionMap::collisionMap;
 
 bool CollisionMap::WalkableTiles(int x, int y) {
     //out of bounds
-    if (y < 0 || y >= collisionMap.size() || x < 0 || x >= collisionMap[0].size()) {
+    if (y < 0 || y >= static_cast<int>(collisionMap.size()) || x < 0 || x >= static_cast<int>(collisionMap[0].size())) {
         return false;
     }
